Add createSelectAnnotation to Annotator
